refactor(folders): update folder structure
